+++ HISTORY.rst
@@ -6,3 +6,16 @@ History
 ------------------
 
 * First release on PyPI.
+
+1.0.1 (2018-10-12)
+------------------
+
+* First production release
+
+
+1.0.2 (2018-10-16)
+------------------
+
+* This patch allows the XMLBarcodeConversion step to also convert inbound
+rule data as opposed to just the data within the context under a specific
+key.